Move Jinja implementation file
